Refactor display widgets to their own folder, optimize some imports.
