+++ cockatrice/src/client/tabs/api/edhrec/api_response/top_commanders/edhrec_top_commanders_api_response.h
@@ -19,4 +19,4 @@ public:
     void debugPrint() const;
 };
 
-#endif //EDHREC_TOP_COMMANDERS_API_RESPONSE_H
+#endif // EDHREC_TOP_COMMANDERS_API_RESPONSE_H
