+++ cockatrice/src/client/tabs/api/edhrec/display/commander/edhrec_commander_api_response_card_details_display_widget.h
@@ -1,8 +1,8 @@
 #ifndef EDHREC_COMMANDER_API_RESPONSE_CARD_DETAILS_DISPLAY_WIDGET_H
 #define EDHREC_COMMANDER_API_RESPONSE_CARD_DETAILS_DISPLAY_WIDGET_H
 
-#include "../../../../ui/widgets/cards/card_info_picture_widget.h"
-#include "../api_response/commander/edhrec_commander_api_response_card_details.h"
+#include "../../../../../ui/widgets/cards/card_info_picture_widget.h"
+#include "../../api_response/commander/edhrec_commander_api_response_card_details.h"
 
 #include <QHBoxLayout>
 #include <QLabel>
