+++ cockatrice/src/client/tabs/api/edhrec/display/commander/edhrec_commander_api_response_card_list_display_widget.h
@@ -1,9 +1,9 @@
 #ifndef EDHREC_COMMANDER_API_RESPONSE_CARD_LIST_DISPLAY_WIDGET_H
 #define EDHREC_COMMANDER_API_RESPONSE_CARD_LIST_DISPLAY_WIDGET_H
 
-#include "../../../../ui/widgets/general/display/banner_widget.h"
-#include "../../../../ui/widgets/general/layout_containers/flow_widget.h"
-#include "../api_response/commander/edhrec_commander_api_response_card_list.h"
+#include "../../../../../ui/widgets/general/display/banner_widget.h"
+#include "../../../../../ui/widgets/general/layout_containers/flow_widget.h"
+#include "../../api_response/commander/edhrec_commander_api_response_card_list.h"
 
 #include <QResizeEvent>
 #include <QVBoxLayout>
