+++ cockatrice/src/client/tabs/api/edhrec/display/commander/edhrec_commander_api_response_display_widget.cpp
@@ -1,7 +1,7 @@
 #include "edhrec_commander_api_response_display_widget.h"
 
-#include "../../../../ui/widgets/cards/card_info_picture_widget.h"
-#include "../api_response/commander/edhrec_commander_api_response.h"
+#include "../../../../../ui/widgets/cards/card_info_picture_widget.h"
+#include "../../api_response/commander/edhrec_commander_api_response.h"
 #include "edhrec_commander_api_response_card_list_display_widget.h"
 #include "edhrec_commander_api_response_commander_details_display_widget.h"
