+++ cockatrice/src/client/tabs/api/edhrec/display/top_cards/edhrec_top_cards_api_response_display_widget.cpp
@@ -1,19 +1,10 @@
-#include "edhrec_top_commanders_api_response_display_widget.h"
+#include "edhrec_top_cards_api_response_display_widget.h"
 
-#include "../../../../ui/widgets/cards/card_info_picture_widget.h"
-#include "../api_response/top_commanders/edhrec_top_commanders_api_response.h"
+#include "../../api_response/top_cards/edhrec_top_cards_api_response.h"
 #include "../commander/edhrec_commander_api_response_card_list_display_widget.h"
-#include "../commander/edhrec_commander_api_response_commander_details_display_widget.h"
 
-#include <QListView>
-#include <QResizeEvent>
-#include <QScrollArea>
-#include <QSplitter>
-#include <QStringListModel>
-
-EdhrecTopCommandersApiResponseDisplayWidget::EdhrecTopCommandersApiResponseDisplayWidget(
-    QWidget *parent,
-    EdhrecTopCommandersApiResponse response)
+EdhrecTopCardsApiResponseDisplayWidget::EdhrecTopCardsApiResponseDisplayWidget(QWidget *parent,
+                                                                               EdhrecTopCardsApiResponse response)
     : QWidget(parent)
 {
     layout = new QHBoxLayout(this);
@@ -42,10 +33,9 @@ EdhrecTopCommandersApiResponseDisplayWidget::EdhrecTopCommandersApiResponseDispl
     layout->addWidget(scrollArea);
 }
 
-void EdhrecTopCommandersApiResponseDisplayWidget::resizeEvent(QResizeEvent *event)
+void EdhrecTopCardsApiResponseDisplayWidget::resizeEvent(QResizeEvent *event)
 {
     QWidget::resizeEvent(event);
-    qDebug() << event->size();
     layout->invalidate();
     layout->activate();
     layout->update();
