+++ cockatrice/src/client/tabs/api/edhrec/display/top_commander/edhrec_top_commanders_api_response_display_widget.cpp
@@ -1,18 +1,11 @@
-#include "edhrec_top_tags_api_response_display_widget.h"
+#include "edhrec_top_commanders_api_response_display_widget.h"
 
-#include "../../../../ui/widgets/cards/card_info_picture_widget.h"
-#include "../api_response/top_commanders/edhrec_top_commanders_api_response.h"
+#include "../../api_response/top_commanders/edhrec_top_commanders_api_response.h"
 #include "../commander/edhrec_commander_api_response_card_list_display_widget.h"
-#include "../commander/edhrec_commander_api_response_commander_details_display_widget.h"
 
-#include <QListView>
-#include <QResizeEvent>
-#include <QScrollArea>
-#include <QSplitter>
-#include <QStringListModel>
-
-EdhrecTopTagsApiResponseDisplayWidget::EdhrecTopTagsApiResponseDisplayWidget(QWidget *parent,
-                                                                             EdhrecTopTagsApiResponse response)
+EdhrecTopCommandersApiResponseDisplayWidget::EdhrecTopCommandersApiResponseDisplayWidget(
+    QWidget *parent,
+    EdhrecTopCommandersApiResponse response)
     : QWidget(parent)
 {
     layout = new QHBoxLayout(this);
@@ -41,9 +34,10 @@ EdhrecTopTagsApiResponseDisplayWidget::EdhrecTopTagsApiResponseDisplayWidget(QWi
     layout->addWidget(scrollArea);
 }
 
-void EdhrecTopTagsApiResponseDisplayWidget::resizeEvent(QResizeEvent *event)
+void EdhrecTopCommandersApiResponseDisplayWidget::resizeEvent(QResizeEvent *event)
 {
     QWidget::resizeEvent(event);
+    qDebug() << event->size();
     layout->invalidate();
     layout->activate();
     layout->update();
