+++ cockatrice/src/client/tabs/api/edhrec/display/top_commander/edhrec_top_commanders_api_response_display_widget.h
@@ -1,8 +1,9 @@
 #ifndef EDHREC_TOP_COMMANDERS_API_RESPONSE_DISPLAY_WIDGET_H
 #define EDHREC_TOP_COMMANDERS_API_RESPONSE_DISPLAY_WIDGET_H
 
-#include "../api_response/top_commanders/edhrec_top_commanders_api_response.h"
+#include "../../api_response/top_commanders/edhrec_top_commanders_api_response.h"
 
+#include <QResizeEvent>
 #include <QScrollArea>
 #include <QVBoxLayout>
 #include <QWidget>
