+++ cockatrice/src/client/tabs/api/edhrec/display/top_tags/edhrec_top_tags_api_response_display_widget.cpp
@@ -1,17 +1,10 @@
-#include "edhrec_top_cards_api_response_display_widget.h"
+#include "edhrec_top_tags_api_response_display_widget.h"
 
-#include "../../../../ui/widgets/cards/card_info_picture_widget.h"
-#include "../api_response/top_cards/edhrec_top_cards_api_response.h"
+#include "../../api_response/top_tags/edhrec_top_tags_api_response.h"
 #include "../commander/edhrec_commander_api_response_card_list_display_widget.h"
 
-#include <QListView>
-#include <QResizeEvent>
-#include <QScrollArea>
-#include <QSplitter>
-#include <QStringListModel>
-
-EdhrecTopCardsApiResponseDisplayWidget::EdhrecTopCardsApiResponseDisplayWidget(QWidget *parent,
-                                                                               EdhrecTopCardsApiResponse response)
+EdhrecTopTagsApiResponseDisplayWidget::EdhrecTopTagsApiResponseDisplayWidget(QWidget *parent,
+                                                                             EdhrecTopTagsApiResponse response)
     : QWidget(parent)
 {
     layout = new QHBoxLayout(this);
@@ -40,7 +33,7 @@ EdhrecTopCardsApiResponseDisplayWidget::EdhrecTopCardsApiResponseDisplayWidget(Q
     layout->addWidget(scrollArea);
 }
 
-void EdhrecTopCardsApiResponseDisplayWidget::resizeEvent(QResizeEvent *event)
+void EdhrecTopTagsApiResponseDisplayWidget::resizeEvent(QResizeEvent *event)
 {
     QWidget::resizeEvent(event);
     layout->invalidate();
