+++ cockatrice/src/client/tabs/api/edhrec/display/top_tags/edhrec_top_tags_api_response_display_widget.h
@@ -1,8 +1,9 @@
 #ifndef EDHREC_TOP_TAGS_API_RESPONSE_DISPLAY_WIDGET_H
 #define EDHREC_TOP_TAGS_API_RESPONSE_DISPLAY_WIDGET_H
 
-#include "../api_response/top_tags/edhrec_top_tags_api_response.h"
+#include "../../api_response/top_tags/edhrec_top_tags_api_response.h"
 
+#include <QResizeEvent>
 #include <QScrollArea>
 #include <QVBoxLayout>
 #include <QWidget>
