+++ cockatrice/src/client/tabs/api/edhrec/tab_edhrec.cpp
@@ -1,7 +1,7 @@
 #include "tab_edhrec.h"
 
 #include "api_response/commander/edhrec_commander_api_response.h"
-#include "commander/edhrec_commander_api_response_display_widget.h"
+#include "display/commander/edhrec_commander_api_response_display_widget.h"
 
 #include <QDebug>
 #include <QHBoxLayout>
