+++ cockatrice/src/client/tabs/api/edhrec/tab_edhrec_main.h
@@ -4,7 +4,7 @@
 #include "../../../../game/cards/card_database.h"
 #include "../../../ui/widgets/general/layout_containers/flow_widget.h"
 #include "../../tab.h"
-#include "commander/edhrec_commander_api_response_display_widget.h"
+#include "display/commander/edhrec_commander_api_response_display_widget.h"
 
 #include <QHBoxLayout>
 #include <QLineEdit>
